Move routes.py into its own component folder
