+++ src/familytoolswap_backend/flasktutorial/__init__.py
@@ -1,0 +1,1 @@
+# This file is required in each folder by python but is intentionally left blank
