+++ start.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python
 from src.familytoolswap_backend import app
 
 app.run()
